cli: dedicated fetch module
This cleans up `main.rs`.
